Remove duplicate FAST component registration and create individual folders for each export containing web component(s) (#114)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This pull request makes the following refactoring changes:

- Creates folders for each individual web component containing export
- Removes extra FAST component initializations that were duplicates

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
Check the site to see that everything is rendering correctly, all FAST components should have been accounted for.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [x] I have updated the project documentation to reflect my changes.
