+++ app/index.tsx
@@ -3,9 +3,11 @@ import { DndProvider } from "react-dnd";
 import HTML5Backend from "react-dnd-html5-backend";
 import ReactDOM from "react-dom";
 import App from "./app";
+import { registerWebComponents } from "./web-components";
 /* eslint-disable-next-line */
 const styles = require("./style.css");
 styles;
+registerWebComponents();
 
 /**
  * Primary render function for app. Called on store updates
