+++ app/web-components/README.md
@@ -1,0 +1,3 @@
+# Web components
+
+This folder contains web components and React components that contain web components and must be interpreted with a pragma file to apply events correctly. The `index.ts` file in the root of this folder should contain any web component registration as well managing the component exports.
